setup for mongo
