+++ .devcontainer/docker-compose.yml
@@ -1,7 +1,7 @@
 version: '3.8'
 
 volumes:
-  couchdb-data:
+  mongo-data:
 
 services:
   app:
@@ -11,24 +11,14 @@ services:
 
     volumes:
       - ../..:/workspaces:cached
-
-    # Overrides default command so things don't shut down after the process ends.
     command: sleep infinity
-
-    # Runs app on the same network as the database container, allows "forwardPorts" in devcontainer.json function.
     network_mode: service:db
 
-    # Use "forwardPorts" in **devcontainer.json** to forward an app port locally. 
-    # (Adding the "ports" property to this file will not forward from a Codespace.)
-
   db:
-    image: couchdb:3
+    image: mongo:8
     restart: unless-stopped
     volumes:
-      - couchdb-data:/opt/couchdb/data
+      - mongo-data:/data/db
     environment:
-      - COUCHDB_USER=invex
-      - COUCHDB_PASSWORD=invex
-
-    # Add "forwardPorts": ["5432"] to **devcontainer.json** to forward PostgreSQL locally.
-    # (Adding the "ports" property to this file will not forward from a Codespace.)
+      - MONGO_INITDB_ROOT_USERNAME=invex
+      - MONGO_INITDB_ROOT_PASSWORD=invex
