+++ invex-server/Cargo.toml
@@ -7,10 +7,11 @@ edition = "2021"
 bevy_reflect = { version = "0.14.2", features = ["uuid"] }
 chrono = { version = "0.4.38", features = ["serde"] }
 convert_case = "0.6.0"
-couch_rs = "0.11.0"
 duration-string = { version = "0.4.0", features = ["serde"] }
 extism = "1.9.0"
 rocket = { version = "0.5.1", features = ["serde_json", "uuid", "json", "tls", "secrets"] }
 serde = { version = "1.0.215", features = ["derive"] }
 serde_json = "1.0.133"
 uuid = { version = "1.11.0", features = ["serde", "v4", "fast-rng"] }
+invex-macros = {path = "invex-macros"}
+mongodb = "3.1.0"
