+++ invex-server/invex-macros/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "invex-macros"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+quote = "1.0.37"
+syn = "2.0.89"
