+++ invex-server/src/util/mod.rs
@@ -3,3 +3,5 @@ use uuid::Uuid;
 pub fn new_id() -> String {
     Uuid::new_v4().to_string()
 }
+
+pub mod database;
